Include JAVA_KEYWORDS in ALL_KEYWORDS
JAVA_KEYWORDS was inadvertently removed from ALL_KEYWORDS in d0fd4e3,
breaking several tests.

This also corrects a test to note a known bug. In the "java" test in
prettify_test.html, we incorrectly highlight "remove" as a keyword,
since that's a C# keyword and the language is not specified.

One preexisting test failure remains after this change: The "dart" test
in prettify_test.html fails due to 0dc62dc.