Support TypeScript, add more JS keywords.
This change adds several "newer" JavaScript keywords to the set of
supported keywords (let, yield, await).

"constructor" is the special name of an ES6 class constructor.

It also defines TypeScript as an alias of JavaScript. This is a bit of a
lie - TypeScript has some syntactical features that are not available in
JavaScript, such as interfaces, but for the most part it matches the
syntax exactly.

I've only added TS keywords that are or were reserved JS words
(abstract, interface, implements, enum) but left out other keywords
TypeScript defines (declare, namespace, module, any, number).

I'm happy to add those in case people think it's a good idea. I think it
would improve TypeScript formatting without inflicting much harm on JS
code.

Specs/references:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Lexical_grammar#Keywords
https://github.com/Microsoft/TypeScript/blob/master/doc/spec.md#2.2.1