refactor the makefile
Instead of relying on timestamped dummy files to track dependencies,
we create a list of separate targets to build from their corresponding
sources. This is a better design because it allows recompiling only
the modified targets instead of all or nothing.

However the timestamp approach was kept for the lang-aliases rule
because the alias targets are not known in advance and have to be
extracted out of the language handler scripts.

The code to extract lang-alias was ported to a node.js script,
instead of the existing solution based on bash + perl + js-interpreter.

The compiled files are placed directly into the "loader" folder, and
and old cruft is dropped (no need for multiple archive formats
TAR/TGZ/BZ2), instead only one ZIP archive is generated in the
"distrib" folder.